requirements.txt
asgiref==3.7.2
boto3==1.20.26
botocore==1.23.54
Django>=4.2,<5.0
dj-database-url==0.5.0
django-ckeditor-5==0.2.10
django-crispy-forms
django-recaptcha==3.0.0
django-jazzmin==3.0.0
django-storages==1.12.3
django-taggit==3.0.0
django-import-export
python-decouple==3.5
whitenoise==5.2.0
shortuuid==1.0.11
django-mathfilters==1.0.0
django-anymail==9.1
requests==2.31.0
gunicorn==21.2.0
channels==4.0.0
crispy-bootstrap5==0.7
environs==9.5.0
psycopg2-binary
stripe
pillow
python-dotenv

